+++ frontend/src/lib/components/settings/GeneralTab.svelte
@@ -83,6 +83,7 @@
     { value: 'breeze-light', label: 'Breeze (Light)' },
     { value: 'catppuccin-latte', label: 'Catppuccin Latte' },
     { value: 'nord-light', label: 'Nord (Light)' },
+    { value: 'pop-light', label: 'Pop! (Light)' },
     { value: 'dark', label: $_('settingsGeneral.themeDark') },
     { value: 'dark-gray', label: $_('settingsGeneral.themeDarkGray') },
     { value: 'dark-balanced', label: $_('settingsGeneral.themeDarkBalanced') },
@@ -91,6 +92,7 @@
     { value: 'catppuccin-mocha', label: 'Catppuccin Mocha' },
     { value: 'dracula', label: 'Dracula' },
     { value: 'nord-dark', label: 'Nord (Dark)' },
+    { value: 'pop-dark', label: 'Pop! (Dark)' },
     { value: 'tokyo-night', label: 'Tokyo Night' },
   ])